fix tiles and completely change the functionality
