+++ .github/workflows/release.yml
@@ -10,42 +10,18 @@ jobs:
     strategy:
       matrix:
         node_version:
-          - "16"
+          - "20"
     steps:
       - name: Setup NodeJs
         uses: actions/setup-node@v2
         with:
           node-version: ${{ matrix.node_version }}
-      - name: Install Depenedencies of node-canvas
-        run: |
-          sudo apt-get install -y build-essential libcairo2-dev \
-          libpango1.0-dev libjpeg-dev libgif-dev librsvg2-dev
-
-      - name: Install Depenedencies of puppeteer
-        run: |
-          sudo apt-get install -y gconf-service libasound2 \
-          libatk1.0-0 libatk-bridge2.0-0 libc6 libcairo2 \
-          libcups2 libdbus-1-3 libexpat1 libfontconfig1 \
-          libgcc1 libgconf-2-4 libgdk-pixbuf2.0-0 libglib2.0-0 \
-          libgtk-3-0 libnspr4 libpango-1.0-0 libpangocairo-1.0-0 \
-          libstdc++6 libx11-6 libx11-xcb1 libxcb1 libxcomposite1 \
-          libxcursor1 libxdamage1 libxext6 libxfixes3 libxi6 \
-          libxrandr2 libxrender1 libxss1 libxtst6 ca-certificates \
-          fonts-liberation libappindicator1 libnss3 lsb-release \
-          xdg-utils wget chromium-browser libgbm-dev
 
       - uses: actions/checkout@v2
-      - name: fully install puppeteer
+
+      - name: Rebuild the webpack
         run: |
-          echo 'RWK_CANCEL_SETUP="true"' > .env
           npm install
-          cd ./node_modules/puppeteer
-          npm run install
-          cd ../..
-      - name: Run full fuzzing
-        run: |
-          echo 'RWK_HEADLESS="true"' > .env
-          npm run setup
       - name: zip
         run: |
           sudo apt install zip
