+++ rwk/kitty.js
@@ -1,0 +1,2 @@
+/*! For license information please see kitty.js.LICENSE.txt */
+(()=>{var t={526:(t,e)=>{"use strict";e.byteLength=function(t){var e=l(t),r=e[0],i=e[1];return 3*(r+i)/4-i},e.toByteArray=function(t){var e,r,o=l(t),s=o[0],a=o[1],f=new n(function(t,e,r){return 3*(e+r)/4-r}(0,s,a)),u=0,h=a>0?s-4:s;for(r=0;r<h;r+=4)e=i[t.charCodeAt(r)]<<18|i[t.charCodeAt(r+1)]<<12|i[t.charCodeAt(r+2)]<<6|i[t.charCodeAt(r+3)],f[u++]=e>>16&255,f[u++]=e>>8&255,f[u++]=255&e;return 2===a&&(e=i[t.charCodeAt(r)]<<2|i[t.charCodeAt(r+1)]>>4,f[u++]=255&e),1===a&&(e=i[t.charCodeAt(r)]<<10|i[t.charCodeAt(r+1)]<<4|i[t.charCodeAt(r+2)]>>2,f[u++]=e>>8&255,f[u++]=255&e),f},e.fromByteArray=function(t){for(var e,i=t.length,n=i%3,o=[],s=16383,l=0,f=i-n;l<f;l+=s)o.push(a(t,l,l+s>f?f:l+s));return 1===n?(e=t[i-1],o.push(r[e>>2]+r[e<<4&63]+"==")):2===n&&(e=(t[i-2]<<8)+t[i-1],o.push(r[e>>10]+r[e>>4&63]+r[e<<2&63]+"=")),o.join("")};for(var r=[],i=[],n="undefined"!=typeof Uint8Array?Uint8Array:Array,o="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",s=0;s<64;++s)r[s]=o[s],i[o.charCodeAt(s)]=s;function l(t){var e=t.length;if(e%4>0)throw new Error("Invalid string. Length must be a multiple of 4");var r=t.indexOf("=");return-1===r&&(r=e),[r,r===e?0:4-r%4]}function a(t,e,i){for(var n,o,s=[],l=e;l<i;l+=3)n=(t[l]<<16&16711680)+(t[l+1]<<8&65280)+(255&t[l+2]),s.push(r[(o=n)>>18&63]+r[o>>12&63]+r[o>>6&63]+r[63&o]);return s.join("")}i["-".charCodeAt(0)]=62,i["_".charCodeAt(0)]=63},206:(t,e,r)=>{"use strict";r.r(e),r.d(e,{bits:()=>n,buffer:()=>u,byte:()=>h,default:()=>g,integer:()=>c,nibble:()=>d,string:()=>p});const i=function(t){for(var e=[],r=0;r<t.length;r++)e[r]=1^t[r];return e},n={and:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=t[i]&e[i];return r},circularShiftLeft:function(t,e){var r=[];if(e>t.length)throw new Error("shift amount can’t be larger than bits array length");for(var i=0;i<t.length;i++)r[(t.length+i-e)%t.length]=t[i];return r},circularShiftRight:function(t,e){var r=[];if(e>t.length)throw new Error("shift amount can’t be larger than bits array length");for(var i=0;i<t.length;i++)r[(i+e)%t.length]=t[i];return r},nand:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=t[i]&e[i]^1;return r},nor:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=1^(t[i]|e[i]);return r},not:i,or:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=t[i]|e[i];return r},reduceAnd:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e&=t[r];return e},reduceNand:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e=e&t[r]^1;return e},reduceNor:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e=1^(e|t[r]);return e},reduceOr:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e|=t[r];return e},reduceXnor:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e^=1^t[r];return e},reduceXor:function(t){if(t.length<2)throw new RangeError("Not enough bits.");for(var e=t[0],r=1;r<t.length;r++)e^=t[r];return e},toBoolean:function(t){for(var e=[],r=0;r<t.length;r++)e[r]=1===t[r];return e},toString:function(t,e,r){if(void 0===e&&(e=0),void 0===r&&(r=" "),!e)return t.join("");for(var i="",n=0;n<t.length;n++)i+="".concat(t[n]),n%e==e-1&&n!==t.length-1&&(i+=r);return i},xnor:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=t[i]^e[i]^1;return r},xor:function(t,e){for(var r=[],i=0;i<t.length;i++)r[i]=t[i]^e[i];return r}},o=function(t){if(!Array.isArray(t)||8!==t.length)throw new RangeError("invalid array length");for(var e=0,r=0;r<8;r++)t[7-r]&&(e|=1<<r);return e},s=function(t){if(t>255||t<0||~~t!==t)throw new RangeError("invalid byte");for(var e=[0,0,0,0,0,0,0,0],r=0;r<8;r++)e[7-r]=t>>r&1;return e},l=function(t,e,r){void 0===e&&(e=0),r||(r=8*t.length-e);var i=Math.floor(e/8),n=Math.floor(r/8)+2,o=[];o.length=8*n;for(var l=0;l<n;l++)if(void 0!==t[i+l]){var a=s(t[i+l]);o[8*l]=a[0],o[8*l+1]=a[1],o[8*l+2]=a[2],o[8*l+3]=a[3],o[8*l+4]=a[4],o[8*l+5]=a[5],o[8*l+6]=a[6],o[8*l+7]=a[7]}var f=e%8;return o.slice(f,f+r)};for(var a=[],f=0;f<32;f++)a[f]=Math.pow(2,f);const u={and:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=t[s]&e[o]}return n},create:function(t){for(var e=[0,0,0,0,0,0,0,0],r=Buffer.alloc(Math.ceil(t.length/8)),i=0;i<r.length;i++){for(var n=0;n<8;n++)t[8*i+n]?e[n]=t[8*i+n]:e[n]=0;r[i]=o(e)}return r},modify:function(t,e,r){void 0===r&&(r=0);for(var i=Math.floor(r/8),n=Math.ceil((r+e.length)/8),s=t.slice(i,n),a=l(s),f=r%8,u=0;u<e.length;u++)a[f++]=e[u];for(var h=n-i,c=0;c<h;c++)s[c]=o(a.slice(8*c,8*(c+1)))},nand:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=~(t[s]&e[o])}return n},nor:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=~(t[s]|e[o])}return n},not:function(t){for(var e=Buffer.alloc(t.length),r=0;r<t.length;r++)e[r]=~t[r];return e},or:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=t[s]|e[o]}return n},read:l,readInt:function(t,e,r){void 0===e&&(e=0),void 0===r&&(r=8);var n=l(t,e,r);if(0===n[0]){for(var o=0,s=0;s<r;s++)n[s]&&(o+=a[r-s-1]);return o}o=-1;var f=i(n);for(s=0;s<r;s++)f[s]&&(o-=a[r-s-1]);return o},readUInt:function(t,e,r){void 0===e&&(e=0),void 0===r&&(r=8);for(var i=l(t,e,r),n=0,o=0;o<r;o++)n+=i[o]*a[r-o-1];return n},xnor:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=t[s]^~e[o]}return n},xor:function(t,e,r){void 0===r&&(r=!1);for(var i=r?e.length:t.length,n=Buffer.alloc(i),o=0;o<i;o++){var s=r?o%t.length:o;n[o]=t[s]^e[o]}return n}},h={read:s,write:o},c={getBit:function(t,e){return t>>e&1},setBit:function(t,e,r){return 1===r?t|1<<e:t&~(1<<e)},toggleBit:function(t,e){return t^1<<e}},d={read:function(t){if(t<16&&t>=0&&Math.floor(t)===t){for(var e=[0,0,0,0],r=0;r<4;r++)e[3-r]=t>>r&1;return e}throw new RangeError("invalid array length")},write:function(t){if(!Array.isArray(t)||4!==t.length)throw new RangeError("invalid array length");for(var e=0,r=0;r<4;r++)t[3-r]&&(e|=1<<r);return e}},p={toBits:function(t){for(var e=[],r=0;r<t.length;r++)"1"===t[r]?e.push(1):"0"===t[r]&&e.push(0);return e}},g={bits:n,buffer:u,byte:h,integer:c,nibble:d,string:p}},287:(t,e,r)=>{"use strict";const i=r(526),n=r(251),o="function"==typeof Symbol&&"function"==typeof Symbol.for?Symbol.for("nodejs.util.inspect.custom"):null;e.hp=a,e.IS=50;const s=2147483647;function l(t){if(t>s)throw new RangeError('The value "'+t+'" is invalid for option "size"');const e=new Uint8Array(t);return Object.setPrototypeOf(e,a.prototype),e}function a(t,e,r){if("number"==typeof t){if("string"==typeof e)throw new TypeError('The "string" argument must be of type string. Received type number');return h(t)}return f(t,e,r)}function f(t,e,r){if("string"==typeof t)return function(t,e){if("string"==typeof e&&""!==e||(e="utf8"),!a.isEncoding(e))throw new TypeError("Unknown encoding: "+e);const r=0|g(t,e);let i=l(r);const n=i.write(t,e);return n!==r&&(i=i.slice(0,n)),i}(t,e);if(ArrayBuffer.isView(t))return function(t){if(Y(t,Uint8Array)){const e=new Uint8Array(t);return d(e.buffer,e.byteOffset,e.byteLength)}return c(t)}(t);if(null==t)throw new TypeError("The first argument must be one of type string, Buffer, ArrayBuffer, Array, or Array-like Object. Received type "+typeof t);if(Y(t,ArrayBuffer)||t&&Y(t.buffer,ArrayBuffer))return d(t,e,r);if("undefined"!=typeof SharedArrayBuffer&&(Y(t,SharedArrayBuffer)||t&&Y(t.buffer,SharedArrayBuffer)))return d(t,e,r);if("number"==typeof t)throw new TypeError('The "value" argument must not be of type number. Received type number');const i=t.valueOf&&t.valueOf();if(null!=i&&i!==t)return a.from(i,e,r);const n=function(t){if(a.isBuffer(t)){const e=0|p(t.length),r=l(e);return 0===r.length||t.copy(r,0,0,e),r}return void 0!==t.length?"number"!=typeof t.length||X(t.length)?l(0):c(t):"Buffer"===t.type&&Array.isArray(t.data)?c(t.data):void 0}(t);if(n)return n;if("undefined"!=typeof Symbol&&null!=Symbol.toPrimitive&&"function"==typeof t[Symbol.toPrimitive])return a.from(t[Symbol.toPrimitive]("string"),e,r);throw new TypeError("The first argument must be one of type string, Buffer, ArrayBuffer, Array, or Array-like Object. Received type "+typeof t)}function u(t){if("number"!=typeof t)throw new TypeError('"size" argument must be of type number');if(t<0)throw new RangeError('The value "'+t+'" is invalid for option "size"')}function h(t){return u(t),l(t<0?0:0|p(t))}function c(t){const e=t.length<0?0:0|p(t.length),r=l(e);for(let i=0;i<e;i+=1)r[i]=255&t[i];return r}function d(t,e,r){if(e<0||t.byteLength<e)throw new RangeError('"offset" is outside of buffer bounds');if(t.byteLength<e+(r||0))throw new RangeError('"length" is outside of buffer bounds');let i;return i=void 0===e&&void 0===r?new Uint8Array(t):void 0===r?new Uint8Array(t,e):new Uint8Array(t,e,r),Object.setPrototypeOf(i,a.prototype),i}function p(t){if(t>=s)throw new RangeError("Attempt to allocate Buffer larger than maximum size: 0x"+s.toString(16)+" bytes");return 0|t}function g(t,e){if(a.isBuffer(t))return t.length;if(ArrayBuffer.isView(t)||Y(t,ArrayBuffer))return t.byteLength;if("string"!=typeof t)throw new TypeError('The "string" argument must be one of type string, Buffer, or ArrayBuffer. Received type '+typeof t);const r=t.length,i=arguments.length>2&&!0===arguments[2];if(!i&&0===r)return 0;let n=!1;for(;;)switch(e){case"ascii":case"latin1":case"binary":return r;case"utf8":case"utf-8":return G(t).length;case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return 2*r;case"hex":return r>>>1;case"base64":return H(t).length;default:if(n)return i?-1:G(t).length;e=(""+e).toLowerCase(),n=!0}}function y(t,e,r){let i=!1;if((void 0===e||e<0)&&(e=0),e>this.length)return"";if((void 0===r||r>this.length)&&(r=this.length),r<=0)return"";if((r>>>=0)<=(e>>>=0))return"";for(t||(t="utf8");;)switch(t){case"hex":return z(this,e,r);case"utf8":case"utf-8":return T(this,e,r);case"ascii":return L(this,e,r);case"latin1":case"binary":return k(this,e,r);case"base64":return B(this,e,r);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return U(this,e,r);default:if(i)throw new TypeError("Unknown encoding: "+t);t=(t+"").toLowerCase(),i=!0}}function m(t,e,r){const i=t[e];t[e]=t[r],t[r]=i}function b(t,e,r,i,n){if(0===t.length)return-1;if("string"==typeof r?(i=r,r=0):r>2147483647?r=2147483647:r<-2147483648&&(r=-2147483648),X(r=+r)&&(r=n?0:t.length-1),r<0&&(r=t.length+r),r>=t.length){if(n)return-1;r=t.length-1}else if(r<0){if(!n)return-1;r=0}if("string"==typeof e&&(e=a.from(e,i)),a.isBuffer(e))return 0===e.length?-1:w(t,e,r,i,n);if("number"==typeof e)return e&=255,"function"==typeof Uint8Array.prototype.indexOf?n?Uint8Array.prototype.indexOf.call(t,e,r):Uint8Array.prototype.lastIndexOf.call(t,e,r):w(t,[e],r,i,n);throw new TypeError("val must be string, number or Buffer")}function w(t,e,r,i,n){let o,s=1,l=t.length,a=e.length;if(void 0!==i&&("ucs2"===(i=String(i).toLowerCase())||"ucs-2"===i||"utf16le"===i||"utf-16le"===i)){if(t.length<2||e.length<2)return-1;s=2,l/=2,a/=2,r/=2}function f(t,e){return 1===s?t[e]:t.readUInt16BE(e*s)}if(n){let i=-1;for(o=r;o<l;o++)if(f(t,o)===f(e,-1===i?0:o-i)){if(-1===i&&(i=o),o-i+1===a)return i*s}else-1!==i&&(o-=o-i),i=-1}else for(r+a>l&&(r=l-a),o=r;o>=0;o--){let r=!0;for(let i=0;i<a;i++)if(f(t,o+i)!==f(e,i)){r=!1;break}if(r)return o}return-1}function v(t,e,r,i){r=Number(r)||0;const n=t.length-r;i?(i=Number(i))>n&&(i=n):i=n;const o=e.length;let s;for(i>o/2&&(i=o/2),s=0;s<i;++s){const i=parseInt(e.substr(2*s,2),16);if(X(i))return s;t[r+s]=i}return s}function E(t,e,r,i){return q(G(e,t.length-r),t,r,i)}function _(t,e,r,i){return q(function(t){const e=[];for(let r=0;r<t.length;++r)e.push(255&t.charCodeAt(r));return e}(e),t,r,i)}function x(t,e,r,i){return q(H(e),t,r,i)}function A(t,e,r,i){return q(function(t,e){let r,i,n;const o=[];for(let s=0;s<t.length&&!((e-=2)<0);++s)r=t.charCodeAt(s),i=r>>8,n=r%256,o.push(n),o.push(i);return o}(e,t.length-r),t,r,i)}function B(t,e,r){return 0===e&&r===t.length?i.fromByteArray(t):i.fromByteArray(t.slice(e,r))}function T(t,e,r){r=Math.min(t.length,r);const i=[];let n=e;for(;n<r;){const e=t[n];let o=null,s=e>239?4:e>223?3:e>191?2:1;if(n+s<=r){let r,i,l,a;switch(s){case 1:e<128&&(o=e);break;case 2:r=t[n+1],128==(192&r)&&(a=(31&e)<<6|63&r,a>127&&(o=a));break;case 3:r=t[n+1],i=t[n+2],128==(192&r)&&128==(192&i)&&(a=(15&e)<<12|(63&r)<<6|63&i,a>2047&&(a<55296||a>57343)&&(o=a));break;case 4:r=t[n+1],i=t[n+2],l=t[n+3],128==(192&r)&&128==(192&i)&&128==(192&l)&&(a=(15&e)<<18|(63&r)<<12|(63&i)<<6|63&l,a>65535&&a<1114112&&(o=a))}}null===o?(o=65533,s=1):o>65535&&(o-=65536,i.push(o>>>10&1023|55296),o=56320|1023&o),i.push(o),n+=s}return function(t){const e=t.length;if(e<=I)return String.fromCharCode.apply(String,t);let r="",i=0;for(;i<e;)r+=String.fromCharCode.apply(String,t.slice(i,i+=I));return r}(i)}a.TYPED_ARRAY_SUPPORT=function(){try{const t=new Uint8Array(1),e={foo:function(){return 42}};return Object.setPrototypeOf(e,Uint8Array.prototype),Object.setPrototypeOf(t,e),42===t.foo()}catch(t){return!1}}(),a.TYPED_ARRAY_SUPPORT||"undefined"==typeof console||"function"!=typeof console.error||console.error("This browser lacks typed array (Uint8Array) support which is required by `buffer` v5.x. Use `buffer` v4.x if you require old browser support."),Object.defineProperty(a.prototype,"parent",{enumerable:!0,get:function(){if(a.isBuffer(this))return this.buffer}}),Object.defineProperty(a.prototype,"offset",{enumerable:!0,get:function(){if(a.isBuffer(this))return this.byteOffset}}),a.poolSize=8192,a.from=function(t,e,r){return f(t,e,r)},Object.setPrototypeOf(a.prototype,Uint8Array.prototype),Object.setPrototypeOf(a,Uint8Array),a.alloc=function(t,e,r){return function(t,e,r){return u(t),t<=0?l(t):void 0!==e?"string"==typeof r?l(t).fill(e,r):l(t).fill(e):l(t)}(t,e,r)},a.allocUnsafe=function(t){return h(t)},a.allocUnsafeSlow=function(t){return h(t)},a.isBuffer=function(t){return null!=t&&!0===t._isBuffer&&t!==a.prototype},a.compare=function(t,e){if(Y(t,Uint8Array)&&(t=a.from(t,t.offset,t.byteLength)),Y(e,Uint8Array)&&(e=a.from(e,e.offset,e.byteLength)),!a.isBuffer(t)||!a.isBuffer(e))throw new TypeError('The "buf1", "buf2" arguments must be one of type Buffer or Uint8Array');if(t===e)return 0;let r=t.length,i=e.length;for(let n=0,o=Math.min(r,i);n<o;++n)if(t[n]!==e[n]){r=t[n],i=e[n];break}return r<i?-1:i<r?1:0},a.isEncoding=function(t){switch(String(t).toLowerCase()){case"hex":case"utf8":case"utf-8":case"ascii":case"latin1":case"binary":case"base64":case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return!0;default:return!1}},a.concat=function(t,e){if(!Array.isArray(t))throw new TypeError('"list" argument must be an Array of Buffers');if(0===t.length)return a.alloc(0);let r;if(void 0===e)for(e=0,r=0;r<t.length;++r)e+=t[r].length;const i=a.allocUnsafe(e);let n=0;for(r=0;r<t.length;++r){let e=t[r];if(Y(e,Uint8Array))n+e.length>i.length?(a.isBuffer(e)||(e=a.from(e)),e.copy(i,n)):Uint8Array.prototype.set.call(i,e,n);else{if(!a.isBuffer(e))throw new TypeError('"list" argument must be an Array of Buffers');e.copy(i,n)}n+=e.length}return i},a.byteLength=g,a.prototype._isBuffer=!0,a.prototype.swap16=function(){const t=this.length;if(t%2!=0)throw new RangeError("Buffer size must be a multiple of 16-bits");for(let e=0;e<t;e+=2)m(this,e,e+1);return this},a.prototype.swap32=function(){const t=this.length;if(t%4!=0)throw new RangeError("Buffer size must be a multiple of 32-bits");for(let e=0;e<t;e+=4)m(this,e,e+3),m(this,e+1,e+2);return this},a.prototype.swap64=function(){const t=this.length;if(t%8!=0)throw new RangeError("Buffer size must be a multiple of 64-bits");for(let e=0;e<t;e+=8)m(this,e,e+7),m(this,e+1,e+6),m(this,e+2,e+5),m(this,e+3,e+4);return this},a.prototype.toString=function(){const t=this.length;return 0===t?"":0===arguments.length?T(this,0,t):y.apply(this,arguments)},a.prototype.toLocaleString=a.prototype.toString,a.prototype.equals=function(t){if(!a.isBuffer(t))throw new TypeError("Argument must be a Buffer");return this===t||0===a.compare(this,t)},a.prototype.inspect=function(){let t="";const r=e.IS;return t=this.toString("hex",0,r).replace(/(.{2})/g,"$1 ").trim(),this.length>r&&(t+=" ... "),"<Buffer "+t+">"},o&&(a.prototype[o]=a.prototype.inspect),a.prototype.compare=function(t,e,r,i,n){if(Y(t,Uint8Array)&&(t=a.from(t,t.offset,t.byteLength)),!a.isBuffer(t))throw new TypeError('The "target" argument must be one of type Buffer or Uint8Array. Received type '+typeof t);if(void 0===e&&(e=0),void 0===r&&(r=t?t.length:0),void 0===i&&(i=0),void 0===n&&(n=this.length),e<0||r>t.length||i<0||n>this.length)throw new RangeError("out of range index");if(i>=n&&e>=r)return 0;if(i>=n)return-1;if(e>=r)return 1;if(this===t)return 0;let o=(n>>>=0)-(i>>>=0),s=(r>>>=0)-(e>>>=0);const l=Math.min(o,s),f=this.slice(i,n),u=t.slice(e,r);for(let t=0;t<l;++t)if(f[t]!==u[t]){o=f[t],s=u[t];break}return o<s?-1:s<o?1:0},a.prototype.includes=function(t,e,r){return-1!==this.indexOf(t,e,r)},a.prototype.indexOf=function(t,e,r){return b(this,t,e,r,!0)},a.prototype.lastIndexOf=function(t,e,r){return b(this,t,e,r,!1)},a.prototype.write=function(t,e,r,i){if(void 0===e)i="utf8",r=this.length,e=0;else if(void 0===r&&"string"==typeof e)i=e,r=this.length,e=0;else{if(!isFinite(e))throw new Error("Buffer.write(string, encoding, offset[, length]) is no longer supported");e>>>=0,isFinite(r)?(r>>>=0,void 0===i&&(i="utf8")):(i=r,r=void 0)}const n=this.length-e;if((void 0===r||r>n)&&(r=n),t.length>0&&(r<0||e<0)||e>this.length)throw new RangeError("Attempt to write outside buffer bounds");i||(i="utf8");let o=!1;for(;;)switch(i){case"hex":return v(this,t,e,r);case"utf8":case"utf-8":return E(this,t,e,r);case"ascii":case"latin1":case"binary":return _(this,t,e,r);case"base64":return x(this,t,e,r);case"ucs2":case"ucs-2":case"utf16le":case"utf-16le":return A(this,t,e,r);default:if(o)throw new TypeError("Unknown encoding: "+i);i=(""+i).toLowerCase(),o=!0}},a.prototype.toJSON=function(){return{type:"Buffer",data:Array.prototype.slice.call(this._arr||this,0)}};const I=4096;function L(t,e,r){let i="";r=Math.min(t.length,r);for(let n=e;n<r;++n)i+=String.fromCharCode(127&t[n]);return i}function k(t,e,r){let i="";r=Math.min(t.length,r);for(let n=e;n<r;++n)i+=String.fromCharCode(t[n]);return i}function z(t,e,r){const i=t.length;(!e||e<0)&&(e=0),(!r||r<0||r>i)&&(r=i);let n="";for(let i=e;i<r;++i)n+=J[t[i]];return n}function U(t,e,r){const i=t.slice(e,r);let n="";for(let t=0;t<i.length-1;t+=2)n+=String.fromCharCode(i[t]+256*i[t+1]);return n}function O(t,e,r){if(t%1!=0||t<0)throw new RangeError("offset is not uint");if(t+e>r)throw new RangeError("Trying to access beyond buffer length")}function R(t,e,r,i,n,o){if(!a.isBuffer(t))throw new TypeError('"buffer" argument must be a Buffer instance');if(e>n||e<o)throw new RangeError('"value" argument is out of bounds');if(r+i>t.length)throw new RangeError("Index out of range")}function M(t,e,r,i,n){D(e,i,n,t,r,7);let o=Number(e&BigInt(4294967295));t[r++]=o,o>>=8,t[r++]=o,o>>=8,t[r++]=o,o>>=8,t[r++]=o;let s=Number(e>>BigInt(32)&BigInt(4294967295));return t[r++]=s,s>>=8,t[r++]=s,s>>=8,t[r++]=s,s>>=8,t[r++]=s,r}function P(t,e,r,i,n){D(e,i,n,t,r,7);let o=Number(e&BigInt(4294967295));t[r+7]=o,o>>=8,t[r+6]=o,o>>=8,t[r+5]=o,o>>=8,t[r+4]=o;let s=Number(e>>BigInt(32)&BigInt(4294967295));return t[r+3]=s,s>>=8,t[r+2]=s,s>>=8,t[r+1]=s,s>>=8,t[r]=s,r+8}function S(t,e,r,i,n,o){if(r+i>t.length)throw new RangeError("Index out of range");if(r<0)throw new RangeError("Index out of range")}function j(t,e,r,i,o){return e=+e,r>>>=0,o||S(t,0,r,4),n.write(t,e,r,i,23,4),r+4}function $(t,e,r,i,o){return e=+e,r>>>=0,o||S(t,0,r,8),n.write(t,e,r,i,52,8),r+8}a.prototype.slice=function(t,e){const r=this.length;(t=~~t)<0?(t+=r)<0&&(t=0):t>r&&(t=r),(e=void 0===e?r:~~e)<0?(e+=r)<0&&(e=0):e>r&&(e=r),e<t&&(e=t);const i=this.subarray(t,e);return Object.setPrototypeOf(i,a.prototype),i},a.prototype.readUintLE=a.prototype.readUIntLE=function(t,e,r){t>>>=0,e>>>=0,r||O(t,e,this.length);let i=this[t],n=1,o=0;for(;++o<e&&(n*=256);)i+=this[t+o]*n;return i},a.prototype.readUintBE=a.prototype.readUIntBE=function(t,e,r){t>>>=0,e>>>=0,r||O(t,e,this.length);let i=this[t+--e],n=1;for(;e>0&&(n*=256);)i+=this[t+--e]*n;return i},a.prototype.readUint8=a.prototype.readUInt8=function(t,e){return t>>>=0,e||O(t,1,this.length),this[t]},a.prototype.readUint16LE=a.prototype.readUInt16LE=function(t,e){return t>>>=0,e||O(t,2,this.length),this[t]|this[t+1]<<8},a.prototype.readUint16BE=a.prototype.readUInt16BE=function(t,e){return t>>>=0,e||O(t,2,this.length),this[t]<<8|this[t+1]},a.prototype.readUint32LE=a.prototype.readUInt32LE=function(t,e){return t>>>=0,e||O(t,4,this.length),(this[t]|this[t+1]<<8|this[t+2]<<16)+16777216*this[t+3]},a.prototype.readUint32BE=a.prototype.readUInt32BE=function(t,e){return t>>>=0,e||O(t,4,this.length),16777216*this[t]+(this[t+1]<<16|this[t+2]<<8|this[t+3])},a.prototype.readBigUInt64LE=Z((function(t){V(t>>>=0,"offset");const e=this[t],r=this[t+7];void 0!==e&&void 0!==r||K(t,this.length-8);const i=e+256*this[++t]+65536*this[++t]+this[++t]*2**24,n=this[++t]+256*this[++t]+65536*this[++t]+r*2**24;return BigInt(i)+(BigInt(n)<<BigInt(32))})),a.prototype.readBigUInt64BE=Z((function(t){V(t>>>=0,"offset");const e=this[t],r=this[t+7];void 0!==e&&void 0!==r||K(t,this.length-8);const i=e*2**24+65536*this[++t]+256*this[++t]+this[++t],n=this[++t]*2**24+65536*this[++t]+256*this[++t]+r;return(BigInt(i)<<BigInt(32))+BigInt(n)})),a.prototype.readIntLE=function(t,e,r){t>>>=0,e>>>=0,r||O(t,e,this.length);let i=this[t],n=1,o=0;for(;++o<e&&(n*=256);)i+=this[t+o]*n;return n*=128,i>=n&&(i-=Math.pow(2,8*e)),i},a.prototype.readIntBE=function(t,e,r){t>>>=0,e>>>=0,r||O(t,e,this.length);let i=e,n=1,o=this[t+--i];for(;i>0&&(n*=256);)o+=this[t+--i]*n;return n*=128,o>=n&&(o-=Math.pow(2,8*e)),o},a.prototype.readInt8=function(t,e){return t>>>=0,e||O(t,1,this.length),128&this[t]?-1*(255-this[t]+1):this[t]},a.prototype.readInt16LE=function(t,e){t>>>=0,e||O(t,2,this.length);const r=this[t]|this[t+1]<<8;return 32768&r?4294901760|r:r},a.prototype.readInt16BE=function(t,e){t>>>=0,e||O(t,2,this.length);const r=this[t+1]|this[t]<<8;return 32768&r?4294901760|r:r},a.prototype.readInt32LE=function(t,e){return t>>>=0,e||O(t,4,this.length),this[t]|this[t+1]<<8|this[t+2]<<16|this[t+3]<<24},a.prototype.readInt32BE=function(t,e){return t>>>=0,e||O(t,4,this.length),this[t]<<24|this[t+1]<<16|this[t+2]<<8|this[t+3]},a.prototype.readBigInt64LE=Z((function(t){V(t>>>=0,"offset");const e=this[t],r=this[t+7];void 0!==e&&void 0!==r||K(t,this.length-8);const i=this[t+4]+256*this[t+5]+65536*this[t+6]+(r<<24);return(BigInt(i)<<BigInt(32))+BigInt(e+256*this[++t]+65536*this[++t]+this[++t]*2**24)})),a.prototype.readBigInt64BE=Z((function(t){V(t>>>=0,"offset");const e=this[t],r=this[t+7];void 0!==e&&void 0!==r||K(t,this.length-8);const i=(e<<24)+65536*this[++t]+256*this[++t]+this[++t];return(BigInt(i)<<BigInt(32))+BigInt(this[++t]*2**24+65536*this[++t]+256*this[++t]+r)})),a.prototype.readFloatLE=function(t,e){return t>>>=0,e||O(t,4,this.length),n.read(this,t,!0,23,4)},a.prototype.readFloatBE=function(t,e){return t>>>=0,e||O(t,4,this.length),n.read(this,t,!1,23,4)},a.prototype.readDoubleLE=function(t,e){return t>>>=0,e||O(t,8,this.length),n.read(this,t,!0,52,8)},a.prototype.readDoubleBE=function(t,e){return t>>>=0,e||O(t,8,this.length),n.read(this,t,!1,52,8)},a.prototype.writeUintLE=a.prototype.writeUIntLE=function(t,e,r,i){t=+t,e>>>=0,r>>>=0,i||R(this,t,e,r,Math.pow(2,8*r)-1,0);let n=1,o=0;for(this[e]=255&t;++o<r&&(n*=256);)this[e+o]=t/n&255;return e+r},a.prototype.writeUintBE=a.prototype.writeUIntBE=function(t,e,r,i){t=+t,e>>>=0,r>>>=0,i||R(this,t,e,r,Math.pow(2,8*r)-1,0);let n=r-1,o=1;for(this[e+n]=255&t;--n>=0&&(o*=256);)this[e+n]=t/o&255;return e+r},a.prototype.writeUint8=a.prototype.writeUInt8=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,1,255,0),this[e]=255&t,e+1},a.prototype.writeUint16LE=a.prototype.writeUInt16LE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,2,65535,0),this[e]=255&t,this[e+1]=t>>>8,e+2},a.prototype.writeUint16BE=a.prototype.writeUInt16BE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,2,65535,0),this[e]=t>>>8,this[e+1]=255&t,e+2},a.prototype.writeUint32LE=a.prototype.writeUInt32LE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,4,4294967295,0),this[e+3]=t>>>24,this[e+2]=t>>>16,this[e+1]=t>>>8,this[e]=255&t,e+4},a.prototype.writeUint32BE=a.prototype.writeUInt32BE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,4,4294967295,0),this[e]=t>>>24,this[e+1]=t>>>16,this[e+2]=t>>>8,this[e+3]=255&t,e+4},a.prototype.writeBigUInt64LE=Z((function(t,e=0){return M(this,t,e,BigInt(0),BigInt("0xffffffffffffffff"))})),a.prototype.writeBigUInt64BE=Z((function(t,e=0){return P(this,t,e,BigInt(0),BigInt("0xffffffffffffffff"))})),a.prototype.writeIntLE=function(t,e,r,i){if(t=+t,e>>>=0,!i){const i=Math.pow(2,8*r-1);R(this,t,e,r,i-1,-i)}let n=0,o=1,s=0;for(this[e]=255&t;++n<r&&(o*=256);)t<0&&0===s&&0!==this[e+n-1]&&(s=1),this[e+n]=(t/o|0)-s&255;return e+r},a.prototype.writeIntBE=function(t,e,r,i){if(t=+t,e>>>=0,!i){const i=Math.pow(2,8*r-1);R(this,t,e,r,i-1,-i)}let n=r-1,o=1,s=0;for(this[e+n]=255&t;--n>=0&&(o*=256);)t<0&&0===s&&0!==this[e+n+1]&&(s=1),this[e+n]=(t/o|0)-s&255;return e+r},a.prototype.writeInt8=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,1,127,-128),t<0&&(t=255+t+1),this[e]=255&t,e+1},a.prototype.writeInt16LE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,2,32767,-32768),this[e]=255&t,this[e+1]=t>>>8,e+2},a.prototype.writeInt16BE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,2,32767,-32768),this[e]=t>>>8,this[e+1]=255&t,e+2},a.prototype.writeInt32LE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,4,2147483647,-2147483648),this[e]=255&t,this[e+1]=t>>>8,this[e+2]=t>>>16,this[e+3]=t>>>24,e+4},a.prototype.writeInt32BE=function(t,e,r){return t=+t,e>>>=0,r||R(this,t,e,4,2147483647,-2147483648),t<0&&(t=4294967295+t+1),this[e]=t>>>24,this[e+1]=t>>>16,this[e+2]=t>>>8,this[e+3]=255&t,e+4},a.prototype.writeBigInt64LE=Z((function(t,e=0){return M(this,t,e,-BigInt("0x8000000000000000"),BigInt("0x7fffffffffffffff"))})),a.prototype.writeBigInt64BE=Z((function(t,e=0){return P(this,t,e,-BigInt("0x8000000000000000"),BigInt("0x7fffffffffffffff"))})),a.prototype.writeFloatLE=function(t,e,r){return j(this,t,e,!0,r)},a.prototype.writeFloatBE=function(t,e,r){return j(this,t,e,!1,r)},a.prototype.writeDoubleLE=function(t,e,r){return $(this,t,e,!0,r)},a.prototype.writeDoubleBE=function(t,e,r){return $(this,t,e,!1,r)},a.prototype.copy=function(t,e,r,i){if(!a.isBuffer(t))throw new TypeError("argument should be a Buffer");if(r||(r=0),i||0===i||(i=this.length),e>=t.length&&(e=t.length),e||(e=0),i>0&&i<r&&(i=r),i===r)return 0;if(0===t.length||0===this.length)return 0;if(e<0)throw new RangeError("targetStart out of bounds");if(r<0||r>=this.length)throw new RangeError("Index out of range");if(i<0)throw new RangeError("sourceEnd out of bounds");i>this.length&&(i=this.length),t.length-e<i-r&&(i=t.length-e+r);const n=i-r;return this===t&&"function"==typeof Uint8Array.prototype.copyWithin?this.copyWithin(e,r,i):Uint8Array.prototype.set.call(t,this.subarray(r,i),e),n},a.prototype.fill=function(t,e,r,i){if("string"==typeof t){if("string"==typeof e?(i=e,e=0,r=this.length):"string"==typeof r&&(i=r,r=this.length),void 0!==i&&"string"!=typeof i)throw new TypeError("encoding must be a string");if("string"==typeof i&&!a.isEncoding(i))throw new TypeError("Unknown encoding: "+i);if(1===t.length){const e=t.charCodeAt(0);("utf8"===i&&e<128||"latin1"===i)&&(t=e)}}else"number"==typeof t?t&=255:"boolean"==typeof t&&(t=Number(t));if(e<0||this.length<e||this.length<r)throw new RangeError("Out of range index");if(r<=e)return this;let n;if(e>>>=0,r=void 0===r?this.length:r>>>0,t||(t=0),"number"==typeof t)for(n=e;n<r;++n)this[n]=t;else{const o=a.isBuffer(t)?t:a.from(t,i),s=o.length;if(0===s)throw new TypeError('The value "'+t+'" is invalid for argument "value"');for(n=0;n<r-e;++n)this[n+e]=o[n%s]}return this};const C={};function F(t,e,r){C[t]=class extends r{constructor(){super(),Object.defineProperty(this,"message",{value:e.apply(this,arguments),writable:!0,configurable:!0}),this.name=`${this.name} [${t}]`,this.stack,delete this.name}get code(){return t}set code(t){Object.defineProperty(this,"code",{configurable:!0,enumerable:!0,value:t,writable:!0})}toString(){return`${this.name} [${t}]: ${this.message}`}}}function N(t){let e="",r=t.length;const i="-"===t[0]?1:0;for(;r>=i+4;r-=3)e=`_${t.slice(r-3,r)}${e}`;return`${t.slice(0,r)}${e}`}function D(t,e,r,i,n,o){if(t>r||t<e){const i="bigint"==typeof e?"n":"";let n;throw n=o>3?0===e||e===BigInt(0)?`>= 0${i} and < 2${i} ** ${8*(o+1)}${i}`:`>= -(2${i} ** ${8*(o+1)-1}${i}) and < 2 ** ${8*(o+1)-1}${i}`:`>= ${e}${i} and <= ${r}${i}`,new C.ERR_OUT_OF_RANGE("value",n,t)}!function(t,e,r){V(e,"offset"),void 0!==t[e]&&void 0!==t[e+r]||K(e,t.length-(r+1))}(i,n,o)}function V(t,e){if("number"!=typeof t)throw new C.ERR_INVALID_ARG_TYPE(e,"number",t)}function K(t,e,r){if(Math.floor(t)!==t)throw V(t,r),new C.ERR_OUT_OF_RANGE(r||"offset","an integer",t);if(e<0)throw new C.ERR_BUFFER_OUT_OF_BOUNDS;throw new C.ERR_OUT_OF_RANGE(r||"offset",`>= ${r?1:0} and <= ${e}`,t)}F("ERR_BUFFER_OUT_OF_BOUNDS",(function(t){return t?`${t} is outside of buffer bounds`:"Attempt to access memory outside buffer bounds"}),RangeError),F("ERR_INVALID_ARG_TYPE",(function(t,e){return`The "${t}" argument must be of type number. Received type ${typeof e}`}),TypeError),F("ERR_OUT_OF_RANGE",(function(t,e,r){let i=`The value of "${t}" is out of range.`,n=r;return Number.isInteger(r)&&Math.abs(r)>2**32?n=N(String(r)):"bigint"==typeof r&&(n=String(r),(r>BigInt(2)**BigInt(32)||r<-(BigInt(2)**BigInt(32)))&&(n=N(n)),n+="n"),i+=` It must be ${e}. Received ${n}`,i}),RangeError);const W=/[^+/0-9A-Za-z-_]/g;function G(t,e){let r;e=e||1/0;const i=t.length;let n=null;const o=[];for(let s=0;s<i;++s){if(r=t.charCodeAt(s),r>55295&&r<57344){if(!n){if(r>56319){(e-=3)>-1&&o.push(239,191,189);continue}if(s+1===i){(e-=3)>-1&&o.push(239,191,189);continue}n=r;continue}if(r<56320){(e-=3)>-1&&o.push(239,191,189),n=r;continue}r=65536+(n-55296<<10|r-56320)}else n&&(e-=3)>-1&&o.push(239,191,189);if(n=null,r<128){if((e-=1)<0)break;o.push(r)}else if(r<2048){if((e-=2)<0)break;o.push(r>>6|192,63&r|128)}else if(r<65536){if((e-=3)<0)break;o.push(r>>12|224,r>>6&63|128,63&r|128)}else{if(!(r<1114112))throw new Error("Invalid code point");if((e-=4)<0)break;o.push(r>>18|240,r>>12&63|128,r>>6&63|128,63&r|128)}}return o}function H(t){return i.toByteArray(function(t){if((t=(t=t.split("=")[0]).trim().replace(W,"")).length<2)return"";for(;t.length%4!=0;)t+="=";return t}(t))}function q(t,e,r,i){let n;for(n=0;n<i&&!(n+r>=e.length||n>=t.length);++n)e[n+r]=t[n];return n}function Y(t,e){return t instanceof e||null!=t&&null!=t.constructor&&null!=t.constructor.name&&t.constructor.name===e.name}function X(t){return t!=t}const J=function(){const t="0123456789abcdef",e=new Array(256);for(let r=0;r<16;++r){const i=16*r;for(let n=0;n<16;++n)e[i+n]=t[r]+t[n]}return e}();function Z(t){return"undefined"==typeof BigInt?Q:t}function Q(){throw new Error("BigInt not supported")}},251:(t,e)=>{e.read=function(t,e,r,i,n){var o,s,l=8*n-i-1,a=(1<<l)-1,f=a>>1,u=-7,h=r?n-1:0,c=r?-1:1,d=t[e+h];for(h+=c,o=d&(1<<-u)-1,d>>=-u,u+=l;u>0;o=256*o+t[e+h],h+=c,u-=8);for(s=o&(1<<-u)-1,o>>=-u,u+=i;u>0;s=256*s+t[e+h],h+=c,u-=8);if(0===o)o=1-f;else{if(o===a)return s?NaN:1/0*(d?-1:1);s+=Math.pow(2,i),o-=f}return(d?-1:1)*s*Math.pow(2,o-i)},e.write=function(t,e,r,i,n,o){var s,l,a,f=8*o-n-1,u=(1<<f)-1,h=u>>1,c=23===n?Math.pow(2,-24)-Math.pow(2,-77):0,d=i?0:o-1,p=i?1:-1,g=e<0||0===e&&1/e<0?1:0;for(e=Math.abs(e),isNaN(e)||e===1/0?(l=isNaN(e)?1:0,s=u):(s=Math.floor(Math.log(e)/Math.LN2),e*(a=Math.pow(2,-s))<1&&(s--,a*=2),(e+=s+h>=1?c/a:c*Math.pow(2,1-h))*a>=2&&(s++,a/=2),s+h>=u?(l=0,s=u):s+h>=1?(l=(e*a-1)*Math.pow(2,n),s+=h):(l=e*Math.pow(2,h-1)*Math.pow(2,n),s=0));n>=8;t[r+d]=255&l,d+=p,l/=256,n-=8);for(s=s<<n|l,f+=n;f>0;t[r+d]=255&s,d+=p,s/=256,f-=8);t[r+d-p]|=128*g}},918:(t,e)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.config=void 0,e.config={debug:!1,base_tiles:{vertical_pairs:[[14,15],[16,17],[18,19],[51,52],[68,69]],horizontal_pairs:[[54,55],[59,60]],parity_pairs:[[38,43]],elevated:[20,11,12,13,50,4,5,6,7,8,9,10,42,72,25,26,65,27,21,22,23,53,56,71,36,24,41,40,31,28,38,43,39,44,45,46,48,57,62,73,70]}}},257:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.register=void 0;const i=r(918),n=r(239),o=r(891);var s=r(287).hp;const l={name:"RwK Level",extension:"kitty",read:function(t){var e;e=`Attempting to open ${t}`,i.config.debug&&tiled.log(e);const r=new BinaryFile(t,BinaryFile.ReadOnly),n=s.from(r.readAll());r.close(),tiled.log("Binary file read to memory.");const l=(0,o.bufToTiled)(n);return tiled.trigger("FitInView"),l},write:function(t,e){const r=e.match(/[^\/]*\.kitty$/),i=r?r[0].slice(0,-6):null;if(!i)throw new Error("Please use a name ending in .kitty.");t.setProperty("name",i),(0,n.fixFlippedCells)(t),tiled.log(`Now exporting level ${i}`);const s=(0,o.tiledToBuf)(t);tiled.log(`Now writing file ${e}`);const l=new BinaryFile(e,BinaryFile.WriteOnly);return l.resize(s.length),l.write(s.buffer),l.commit(),""}};e.register=function(){tiled.log("registering kitty map export format"),tiled.registerMapFormat("kitty",l)}},967:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.register=void 0;const i=r(918),n=r(891);var o=r(287).hp;const s={name:"RwK Level in bas64",extension:"kitty.b64",read:function(t){var e;e=`Attempting to open ${t}`,i.config.debug&&tiled.log(e);const r=new TextFile(t,TextFile.ReadOnly),s=o.from(o.from(r.readAll(),"base64"));r.close(),tiled.log("Buffer file read to memory.");const l=(0,n.bufToTiled)(s);return tiled.trigger("FitInView"),l}};e.register=function(){tiled.log("registering kittyb64 map import format"),tiled.registerMapFormat("kitty.b64",s)}},586:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.register=void 0,r(918);const i=r(239),n=r(891);r(287).hp;const o={name:"RwK Level injection snippet",extension:"kitty.js",write:function(t,e){(0,i.fixFlippedCells)(t);const r=t.property("name");tiled.log(`Now exporting level ${r}`);const o=`\nvar dbRequest = indexedDB.open("/RAPTISOFT_SANDBOX", 21);\ndbRequest.onsuccess = function(event) {\n  var db = event.target.result;\n\n  // Perform the add operation with the key\n  var request = db.transaction(["FILE_DATA"], "readwrite").objectStore("FILE_DATA").put(\n      {\n        timestamp:new Date(),\n        mode: 33206,\n        contents:base64ToUint8Array("${(0,n.tiledToBuf)(t).toString("base64")}")\n      },\n      "/RAPTISOFT_SANDBOX/RWK/EXTRALEVELS64/${r}.kitty");\n\n  request.onsuccess = function() {\n    console.log("Level saved successfully! Reloading in 2 seconds.");\n    setTimeout(function() {location.reload();}, 2000);\n  };\n\n  request.onerror = function(event) {\n    console.error("Error storing Uint8Array data: ", event.target.error);\n  };\n};\n\ndbRequest.onerror = function(event) {\n  console.error("Error opening database: ", event.target.error);\n};\n\nfunction base64ToUint8Array(base64) {\n  return new Uint8Array(Array.from(atob(base64), c => c.charCodeAt(0)));\n}\n        `;tiled.log(`Now writing file ${e}`);const s=new TextFile(e,TextFile.WriteOnly);return s.write(o),s.commit(),""}};e.register=function(){tiled.log("registering kittyjs map export format"),tiled.registerMapFormat("kitty.js",o)}},73:function(t,e,r){"use strict";var i=this&&this.__createBinding||(Object.create?function(t,e,r,i){void 0===i&&(i=r);var n=Object.getOwnPropertyDescriptor(e,r);n&&!("get"in n?!e.__esModule:n.writable||n.configurable)||(n={enumerable:!0,get:function(){return e[r]}}),Object.defineProperty(t,i,n)}:function(t,e,r,i){void 0===i&&(i=r),t[i]=e[r]}),n=this&&this.__setModuleDefault||(Object.create?function(t,e){Object.defineProperty(t,"default",{enumerable:!0,value:e})}:function(t,e){t.default=e}),o=this&&this.__importStar||function(t){if(t&&t.__esModule)return t;var e={};if(null!=t)for(var r in t)"default"!==r&&Object.prototype.hasOwnProperty.call(t,r)&&i(e,t,r);return n(e,t),e};Object.defineProperty(e,"__esModule",{value:!0});const s=o(r(870)),l=o(r(257)),a=o(r(586)),f=o(r(967)),u=o(r(469));tiled.log("rwklevelfiles by MrMasterplan extensions for Tiled"),tiled.log(__filename),s.register(),l.register(),a.register(),f.register(),u.register()},239:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.fixFlippedCells=void 0;const i=r(891),n=r(870),o=r(918);function s(t){if(8!=t.length)throw new Error(`attempting to flip grid of invalid size ${t.length}`);return[t.charAt(2),t.charAt(1),t.charAt(0),t.charAt(4),t.charAt(3),t.charAt(7),t.charAt(6),t.charAt(5)].join("")}function l(t){if(8!=t.length)throw new Error(`attempting to flip grid of invalid size ${t.length}`);return[t.charAt(5),t.charAt(6),t.charAt(7),t.charAt(3),t.charAt(4),t.charAt(0),t.charAt(1),t.charAt(2)].join("")}function a(t){if(8!=t.length)throw new Error(`attempting to flip grid of invalid size ${t.length}`);return[t.charAt(7),t.charAt(4),t.charAt(2),t.charAt(6),t.charAt(1),t.charAt(5),t.charAt(3),t.charAt(0)].join("")}e.fixFlippedCells=function(t){const e=(0,i.getMapLimits)(t),r=[];for(let i=0;i<t.layerCount;i++){const f=t.layerAt(i);if(f.isTileLayer){const t=f;for(let i=e.minx;i<e.maxx;i++){e.minx;for(let f=e.miny;f<e.maxy;f++){e.miny;const u=t.flagsAt(i,f);if(!u)continue;let h=t.tileAt(i,f);if(!h)continue;const c=h.resolvedProperties();if(c.paint){if(!c.paint_grid){tiled.log(`Error in layer ${t.name} at (${i},${f}) tile has paint but not paint_grid property.`);continue}let e=c.paint_grid.toString();if(u&Tile.FlippedHorizontally&&(e=s(e)),u&Tile.FlippedVertically&&(e=l(e)),u&Tile.FlippedAntiDiagonally&&(e=a(e)),e!=c.paint_grid){const t=(0,n.parsed_tileset)(h.tileset)[e];if(!t)throw new Error(`invalid paint grid ${e}`);h=t}r.push((()=>{const e=t.edit();e.setTile(i,f,h,0),e.apply()}))}else if(c.base||c.map){const e=c.base||c.map,s=(0,n.parsed_base_tileset)(h.tileset);let l=!1;if(u&Tile.FlippedVertically)for(let n of o.config.base_tiles.vertical_pairs)if(n.includes(e)){const o=s[n[n.indexOf(e)?0:1]];r.push((()=>{const e=t.edit();e.setTile(i,f,o,0),e.apply()})),l=!0;break}if(!l&&u&Tile.FlippedHorizontally)for(let n of o.config.base_tiles.horizontal_pairs)if(n.includes(e)){const o=s[n[n.indexOf(e)?0:1]];r.push((()=>{const e=t.edit();e.setTile(i,f,o,0),e.apply()})),l=!0;break}if(!l&&!!(u&Tile.FlippedHorizontally)!=!!(u&Tile.FlippedVertically)!=!!(u&Tile.FlippedAntiDiagonally))for(let n of o.config.base_tiles.parity_pairs)if(n.includes(e)){const o=s[n[n.indexOf(e)?0:1]];r.push((()=>{const e=t.edit();e.setTile(i,f,o,0),e.apply()})),l=!0;break}l||r.push((()=>{const e=t.edit();e.setTile(i,f,h,0),e.apply()}))}}}}}t.macro("fix flipped tiles",(()=>{for(let t of r)t()}))}},469:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.register=void 0;const i=r(67),n=r(239);e.register=function(){tiled.log("Now registering Kitty menu tool.");{const t=tiled.registerAction("OpenKittyTilesets",(function(t){(0,i.allPaintTilesets)(),(0,i.mapTileset)(),(0,i.robotTileset)(),(0,i.baseTileset)()}));t.text="Open Kitty Tilesets",t.checkable=!1,t.iconVisibleInMenu=!0,t.icon="rwk/tiles/kitty.png",tiled.extendMenu("Edit",[{action:"OpenKittyTilesets",before:"SelectAll"},{separator:!0}])}{const t=tiled.registerAction("FixKittyMap",(function(t){var e;if(!(null===(e=tiled.activeAsset)||void 0===e?void 0:e.isTileMap))throw new Error("Action is only applicable to maps.");(0,n.fixFlippedCells)(tiled.activeAsset)}));t.text="Fix Kitty Map",t.checkable=!1,t.iconVisibleInMenu=!0,t.icon="rwk/tiles/kitty.png",tiled.extendMenu("Edit",[{action:"FixKittyMap",before:"OpenKittyTilesets"}])}{const t=r(330),e=tiled.registerAction("AboutKitty",(function(e){tiled.alert(`Extension for Robot Wants Kitty level files\nby ${t.author}\nversion ${t.version}\nhomepage ${t.homepage}`,"About Kitty")}));e.text="About .kitty extension",e.checkable=!1,e.iconVisibleInMenu=!0,e.icon="rwk/tiles/kitty.png",tiled.extendMenu("Help",[{action:"AboutKitty"}])}}},67:(t,e)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.mapTileset=e.baseTileset=e.allPaintTilesets=e.robotTileset=void 0,e.robotTileset=function(){const t=tiled.open("ext:/rwk/tilesets/robot.json");if(!t)throw new Error("Unable to open robot tileset");return t},e.allPaintTilesets=function(){const t=["ext:/rwk/tilesets/acid.json","ext:/rwk/tilesets/alu.json","ext:/rwk/tilesets/grate.json","ext:/rwk/tilesets/lava.json","ext:/rwk/tilesets/mario.json","ext:/rwk/tilesets/mesh.json","ext:/rwk/tilesets/panels.json","ext:/rwk/tilesets/steel.json","ext:/rwk/tilesets/velcro.json","ext:/rwk/tilesets/warn.json"],e=[];for(let r of t){const t=tiled.open(r);if(!t)throw new Error(`Unable to open tileset ${r}`);e.push(t)}return e},e.baseTileset=function(){const t=tiled.open("ext:/rwk/tilesets/base.json");if(!t)throw new Error("Unable to open base tileset");return t},e.mapTileset=function(){const t=tiled.open("ext:/rwk/tilesets/map.json");if(!t)throw new Error("Unable to open map tileset");return t}},870:(t,e)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.register=e.cleanPaintGrid=e.parsed_base_tileset=e.parsed_tileset=void 0;const r={};function i(t){let e={};if(e=tiled.__parsed_kitty_tilesets||{},!e[t.name]){e[t.name]={};for(let r of t.tiles){if(!r)continue;const i=r.resolvedProperty("paint_grid");i&&(e[t.name][n(i.toString())]=r)}tiled.__parsed_kitty_tilesets=e}return e[t.name]}function n(t){const e=t.split("").map((t=>+t));return e[1]&&e[3]||(e[0]=0),e[1]&&e[4]||(e[2]=0),e[6]&&e[3]||(e[5]=0),e[6]&&e[4]||(e[7]=0),e.map((t=>t.toString())).join("")}r.name="kittypaint",r.icon="rwk/tiles/80be8000.png",r.activated=function(){this.isPaintable()},r.isPaintable=()=>{var t;const e=!!(null===(t=tiled.mapEditor.tilesetsView.currentTileset)||void 0===t?void 0:t.property("is_paintable"));return e||(tiled.alert("Current tileset is not a paint tileset. Kittypaint will not work.","Warning"),tiled.trigger("StampTool")),e},e.parsed_tileset=i,e.parsed_base_tileset=function(t){let e={};if(e=tiled.__parsed_kitty_tilesets||{},!e[t.name]){e[t.name]={};for(let r of t.tiles){if(!r)continue;const i=r.resolvedProperty("base")||r.resolvedProperty("map");i&&(e[t.name][i]=r)}tiled.__parsed_kitty_tilesets=e}return e[t.name]},e.cleanPaintGrid=n,r.layer=function(){const t=this.map.currentLayer;if(!t.isTileLayer||!("tileAt"in t))throw tiled.alert("can only use kitty paint tool on tile layer."),new Error("can only use kitty paint tool on tile layer.");return t},r.tilePositionChanged=function(){this.pressed&&this.executeOnCurrentPos()},r.mousePressed=function(){this.isPaintable()&&(this.pressed=!0,this.executeOnCurrentPos())},r.mouseReleased=function(){this.pressed=!1},r.executeOnCurrentPos=function(){const{x:t,y:e}=this.tilePosition;if(!this.isPaintable())throw new Error("Error in script.");const r=tiled.mapEditor.tilesetsView.currentTileset;this.updateTileAt(t,e);for(let i=t-1;i<t+2;i+=1)for(let n=e-1;n<e+2;n+=1){if(i==t&&n==e)continue;const o=this.layer().tileAt(i,n);o&&(o.tileset.name===r.name&&this.updateTileAt(i,n))}},r.updateTileAt=function(t,e){var r,o,s,l,a,f,u,h;const c=tiled.mapEditor.tilesetsView.currentTileset;let d=[];d.push((null===(r=this.layer().tileAt(t-1,e-1))||void 0===r?void 0:r.tileset.name)==c.name),d.push((null===(o=this.layer().tileAt(t+0,e-1))||void 0===o?void 0:o.tileset.name)==c.name),d.push((null===(s=this.layer().tileAt(t+1,e-1))||void 0===s?void 0:s.tileset.name)==c.name),d.push((null===(l=this.layer().tileAt(t-1,e))||void 0===l?void 0:l.tileset.name)==c.name),d.push((null===(a=this.layer().tileAt(t+1,e))||void 0===a?void 0:a.tileset.name)==c.name),d.push((null===(f=this.layer().tileAt(t-1,e+1))||void 0===f?void 0:f.tileset.name)==c.name),d.push((null===(u=this.layer().tileAt(t+0,e+1))||void 0===u?void 0:u.tileset.name)==c.name),d.push((null===(h=this.layer().tileAt(t+1,e+1))||void 0===h?void 0:h.tileset.name)==c.name);const p=n(d.map((t=>(+t).toString())).join("")),g=this.layer().tileAt(t,e);if(p===(null==g?void 0:g.resolvedProperty("paint_grid"))&&(null==g?void 0:g.tileset.name)===c.name)return;const y=i(c)[p],m=this.layer().edit();m.setTile(t,e,y),m.apply()},e.register=function(){tiled.log("Now registering Kitty Paint tool."),tiled.registerTool(r.name,r)}},844:function(t,e,r){"use strict";var i=this&&this.__importDefault||function(t){return t&&t.__esModule?t:{default:t}};Object.defineProperty(e,"__esModule",{value:!0}),e.extractLevelName=e.normalizedTileValue=e.Level=void 0;const n=i(r(206));var o=r(287).hp;const s={header_1:o.from("0000000005000000","hex"),header_2:o.from("ffffffff","hex"),footer_head:o.from("0100000000010000000000000000000000000000000000000000000000006666663f0000000000000000b800bc000000000000000000000000000000","hex"),footer:o.from("0100000000010000000000000000000000000000000000000000000000006666663f0000000000000000b800bc0000000000000000000000000000000000003f0000003f010000000001000000000100000000","hex")};function l(t){const e=o.alloc(t.length+4);return e.writeUInt32LE(t.length),t.copy(e,4),e}class a{constructor(){this.length=19,this.tag_names_BE=["kids","easy","hard","insane","tricky","silly","arcade","prank","unfair","evil","rpg","tiny","huge","glitchy"],this.tag_names=["prank","arcade","silly","tricky","insane","hard","easy","kids",void 0,void 0,"glitchy","huge","tiny","rpg","evil","unfair"],this.buf=o.alloc(this.length,0),this.tag_index={};for(let t=0;t<this.tag_names.length;t++){const e=this.tag_names[t];"string"==typeof e&&(this.tag_index[e]=t)}}deserialize(t,e){return this.buf=t.slice(e,e+19),e+19}set_tag(t){const e=this.tag_index[t];if(void 0===e)throw Error(`Given tag "${t}" is unknown.`);const r=n.default.buffer.read(this.buf,0,this.tag_names.length);r[e]=1,n.default.buffer.modify(this.buf,r)}get_tags(){const t=n.default.buffer.read(this.buf,0,this.tag_names.length);let e=[];for(let r=0;r<this.tag_names.length;r++)if(t[r]){const t=this.tag_names[r];t&&e.push(t)}return e.join(",")}serialize(){return this.buf}}class f{constructor(t){t?"val"in t?this._val=t.val:(this._val=0,this.paint=t.paint,this.base=t.base):this._val=0}_normalizedCellValue(t){if(t<65536)return t;const e=o.alloc(4);return e.writeUInt32LE(t),o.from("8000","hex").copy(e,2),e.readUInt32LE()}_normalizedPaint(t){return 4294967168&this._normalizedCellValue(t)}_normalizedBase(t){return 127&this._normalizedCellValue(t)}get value(){return this._val}set value(t){this._val=this._normalizedCellValue(t)}get base(){return this._normalizedBase(this._val)}get paint(){return this._normalizedPaint(this._val)}set base(t){this._val=this.paint+this._normalizedBase(t)}set paint(t){this._val=this.base+this._normalizedPaint(t)}}class u{constructor(t=10,e=7,r=0){this.size_x=0,this.size_y=0,this.celldata=[],this.mapgrid=[],this.setSize(t,e,r)}getCellID(t,e){return e*this.size_x+t}serialize(){const t=o.alloc(this.size_x*this.size_y*5+8);let e=0;t.writeUInt32LE(this.size_x,e),e+=4,t.writeUInt32LE(this.size_y,e),e+=4;for(let r=0;r<this.size_y;r++)for(let i=0;i<this.size_x;i++)t.writeUInt32LE(m(this.celldata[r][i]),e),e+=4;for(let r=0;r<this.size_y;r++)for(let i=0;i<this.size_x;i++)t.writeUInt8(this.mapgrid[r][i],e),e+=1;return l(t)}deserialize(t,e){const r=t.readUInt32LE(e);e+=4;const i=t.readUInt32LE(e);e+=4;const n=t.readUInt32LE(e);if(e+=4,r!=i*n*5+8)throw new Error("Grid size mismatch detected.");this.setSize(i,n,0);for(let r=0;r<this.size_y;r++)for(let i=0;i<this.size_x;i++)this.setCell(i,r,m(t.readUInt32LE(e))),e+=4;for(let r=0;r<this.size_y;r++)for(let i=0;i<this.size_x;i++)this.setMapCell(i,r,t.readUInt8(e)),e+=1;return e}setSize(t=10,e=7,r=0){this.size_x=t,this.size_y=e,this.celldata=Array.from(Array(e),(e=>Array.from(Array(t),(t=>r)))),this.mapgrid=Array.from(Array(e),(e=>Array.from(Array(t),(t=>0))))}_checkLimits(t,e){if(t>=this.size_x||e>=this.size_y||t<0||e<0)throw new Error("invalid grid index")}setCell(t,e,r){this._checkLimits(t,e),this.celldata[e][t]=m(r)}setCellBase(t,e,r){this._checkLimits(t,e);const i=new f({val:this.celldata[e][t]});i.base=r,this.celldata[e][t]=i.value}setCellPaint(t,e,r){this._checkLimits(t,e);const i=new f({val:this.celldata[e][t]});i.paint=r,this.celldata[e][t]=i.value}getCell(t,e){return this._checkLimits(t,e),m(this.celldata[e][t])}getCellObj(t,e){return this._checkLimits(t,e),new f({val:this.celldata[e][t]})}setMapCell(t,e,r){this._checkLimits(t,e),this.mapgrid[e][t]=r}getMapCell(t,e){return this._checkLimits(t,e),this.mapgrid[e][t]}}class h{constructor(){this.callouts=[]}addCallout(t,e,r){this.callouts.push({x:t,y:e,text:r})}serialize(){const t=[];for(let e of this.callouts){const r=o.from(e.text,"utf-8"),i=o.alloc(r.length+13,0);i.writeUInt32LE(e.x,0),i.writeUInt32LE(e.y,4),i.writeUInt32LE(r.length+1,8),r.copy(i,12),t.push(i)}const e=o.concat(t),r=o.alloc(4);return r.writeUInt32LE(this.callouts.length),l(o.concat([r,e]))}deserialize(t,e){this.callouts=[];const r=e+4+t.readUInt32LE(e);e+=4;const i=t.readUInt32LE(e);e+=4;for(let r=0;r<i;r++){const r=t.readUInt32LE(e);e+=4;const i=t.readUInt32LE(e);e+=4;const n=t.readUInt32LE(e);e+=4;const o=t.slice(e,e+n-1).toString("utf-8");e+=n,this.callouts.push({x:r,y:i,text:o})}if(e!=r)throw new Error("Deserialization problem in callout section. Unexpected final size.");return e}}class c{constructor(t,e){[this.x,this.y]=[t,e]}serialize(){const t=o.alloc(8);return t.writeFloatLE(this.x,0),t.writeFloatLE(this.y,4),l(t)}deserialize(t,e){return t.readUInt32LE(e),e+=4,this.x=t.readFloatLE(e),e+=4,this.y=t.readFloatLE(e),e+4}}class d{constructor(){this.header=s.footer_head,this.belt_speed=.6,this.music=["","",""]}serialize(){const t=[];t.push(s.footer_head);const e=o.alloc(4,0);e.writeFloatLE(this.belt_speed),t.push(e),t.push(e);for(let e=0;e<3;e++)t.push(l(o.from(this.music[e]+"\0","utf-8")));return l(o.concat(t))}deserialize(t,e){const r=e+4+t.readUInt32LE(e);e+=4,this.header=o.from(t.slice(e,e+s.footer_head.length)),e+=s.footer_head.length,this.belt_speed=t.readFloatLE(e),e+=8;for(let i=0;i<3;i++){if(e==r)return e;const n=t.readUInt32LE(e);e+=4,this.music[i]=t.slice(e,e+n-1).toString("utf-8"),e+=n}if(e!=r)throw new Error("Final offset in footer is not as expected.");return e}}class p{constructor(){this.name=""}serialize(){const t=o.from(this.name,"utf-8"),e=[];e.push(s.header_1),e.push(new g(4+s.header_1.length+4+s.header_2.length+4+t.length).serialize()),e.push(s.header_2);const r=o.alloc(5+t.length,0);return r.writeUInt32LE(t.length+1),t.copy(r,4),o.concat([l(o.concat(e)),r])}deserialize(t,e){e+=4+t.readUInt32LE(e);const r=t.readUInt32LE(e);return e+=4,this.name=t.slice(e,e+r-1).toString("utf-8"),e+r}}class g{constructor(t){this.val=t,this.buf=o.alloc(4,0),this.buf.writeUInt32LE(t)}serialize(){return this.buf}deserialize(t,e){const r=t.readUInt32LE(e);if(r!=this.val)throw new Error(`Deserialization. Unexpected value. Expected ${this.val}, got ${r}`);return e+4}}class y{constructor(t){var e,r,i,n,o,s,l;this.header=new p,this.header.name=(null==t?void 0:t.name)||"",this.tags=new a,this.grid=new u(null===(e=null==t?void 0:t.grid)||void 0===e?void 0:e.x,null===(r=null==t?void 0:t.grid)||void 0===r?void 0:r.y,null===(i=null==t?void 0:t.grid)||void 0===i?void 0:i.val),this.callouts=new h,this.robot=new c((null===(n=null==t?void 0:t.robot)||void 0===n?void 0:n.x)||50,(null===(o=null==t?void 0:t.robot)||void 0===o?void 0:o.y)||60),this.kitty=new c((null===(s=null==t?void 0:t.kitty)||void 0===s?void 0:s.x)||70,(null===(l=null==t?void 0:t.kitty)||void 0===l?void 0:l.y)||80),this.footer=new d,this.zero=new g(0),this.one=new g(1),this.debug=null==t?void 0:t.debug}set name(t){this.header.name=t}get name(){return this.header.name}static from(t,e){const r=new y({debug:e});return r.deserialize(t),r}changeGridSize(t,e){this.grid=new u(t,e)}dbg(t){this.debug&&this.debug(t)}serialize(){const t=[];return t.push(this.header.serialize()),t.push(this.tags.serialize()),t.push(this.grid.serialize()),t.push(this.one.serialize()),t.push(this.callouts.serialize()),t.push(this.zero.serialize()),t.push(this.robot.serialize()),t.push(this.zero.serialize()),t.push(this.kitty.serialize()),t.push(this.zero.serialize()),t.push(this.footer.serialize()),t.push(this.zero.serialize()),t.push(this.zero.serialize()),o.concat(t)}deserialize(t){this.dbg("beginning level deserialization");let e=this.header.deserialize(t,0);this.dbg("header deser passed"),e=this.tags.deserialize(t,e),this.dbg("tags deser passed"),e=this.grid.deserialize(t,e),this.dbg("grid deser passed"),e=this.one.deserialize(t,e),e=this.callouts.deserialize(t,e),this.dbg("callouts deser passed"),e=this.zero.deserialize(t,e),e=this.robot.deserialize(t,e),e=this.zero.deserialize(t,e),e=this.kitty.deserialize(t,e),e=this.zero.deserialize(t,e),this.dbg("robot and kitty deser passed"),e=this.footer.deserialize(t,e),this.dbg("footer deser passed"),e=this.zero.deserialize(t,e),e=this.zero.deserialize(t,e),this.dbg("complete deserilaization passed")}}function m(t){if(t<65536)return t;const e=o.alloc(4);return e.writeUInt32LE(t),o.from("8000","hex").copy(e,2),e.readUInt32LE()}e.Level=y,e.normalizedTileValue=m,e.extractLevelName=function(t){let e=0;e+=4+t.readUInt32LE(e);const r=t.readUInt32LE(e);return e+=4,t.slice(e,e+r-1).toString("utf-8").trim()}},891:(t,e,r)=>{"use strict";Object.defineProperty(e,"__esModule",{value:!0}),e.getMapLimits=e.tiledToBuf=e.bufToTiled=void 0;const i=r(844),n=r(918),o=r(67),s=r(239);function l(t){n.config.debug&&tiled.log(t)}function a(t){const e={minx:0,miny:0,maxx:0,maxy:0};let r=!1;for(let i=0;i<t.layerCount;i++){const n=t.layerAt(i);if(!n.isTileLayer)continue;const o=n.region();r?(e.minx=Math.min(e.minx,o.boundingRect.x),e.miny=Math.min(e.miny,o.boundingRect.y),e.maxx=Math.max(e.maxx,o.boundingRect.x+o.boundingRect.width),e.maxy=Math.max(e.maxy,o.boundingRect.y+o.boundingRect.height)):(r=!0,e.minx=o.boundingRect.x,e.miny=o.boundingRect.y,e.maxx=o.boundingRect.x+o.boundingRect.width,e.maxy=o.boundingRect.y+o.boundingRect.height)}return e}r(287).hp,e.bufToTiled=function(t){const e=new TileMap;e.setTileSize(40,40),e.infinite=!0,l("Read operation started"),l("Parse all tilesets");for(let t of[(0,o.baseTileset)(),(0,o.mapTileset)(),(0,o.robotTileset)()].concat((0,o.allPaintTilesets)()))e.addTileset(t),tiled.close(t);const r={},s={},a={},f={};for(let t of e.tilesets)for(let e of t.tiles){const n=e.properties();if(n.paint)r[(0,i.normalizedTileValue)(+n.paint)]=e;else if(n.base)s[(0,i.normalizedTileValue)(+n.base)]=e;else if(n.map)a[(0,i.normalizedTileValue)(+n.map)]=e;else{if(!n.kind)throw new Error(`Tile of unknown type in tileset ${t.name}`);if("robot"===n.kind)f.robot=e;else{if("kitty"!==n.kind)throw new Error(`tile in robot tileset has invalid kind property ${n.kind}`);f.kitty=e}}}if(!f.robot||!f.kitty)throw new Error("robot and/or kitty tiles not found");const u=i.Level.from(t,l);tiled.log(`Level parsed. Name: ${u.name}`),e.setProperty("name",u.name),e.setProperty("tags",u.tags.get_tags()),e.setProperty("conveyor_speed",u.footer.belt_speed),e.setProperty("music_red",u.footer.music[0]),e.setProperty("music_green",u.footer.music[1]),e.setProperty("music_blue",u.footer.music[2]),e.setSize(u.grid.size_x,u.grid.size_y);const h=new TileLayer("map");e.addLayer(h);const c=new TileLayer("base");e.addLayer(c);const d=new TileLayer("paint");e.addLayer(d);const p=new TileLayer("items");e.addLayer(p);const g=c.edit(),y=h.edit(),m=d.edit(),b=p.edit();for(let t=0;t<u.grid.size_y;t++)for(let e=0;e<u.grid.size_x;e++){const i=u.grid.getCellObj(e,t);m.setTile(e,t,r[i.paint]),y.setTile(e,t,a[u.grid.getMapCell(e,t)]),n.config.base_tiles.elevated.includes(i.base)?b.setTile(e,t,s[i.base]):g.setTile(e,t,s[i.base])}g.apply(),b.apply(),m.apply(),y.apply();const w=new MapObject;w.tile=f.robot,w.pos.x=u.robot.x-e.tileWidth/2,w.pos.y=u.robot.y+e.tileHeight/2,w.width=f.robot.width,w.height=f.robot.height;const v=new MapObject;v.tile=f.kitty,v.pos.x=u.kitty.x-e.tileWidth/2,v.pos.y=u.kitty.y+e.tileHeight/2,v.width=f.kitty.width,v.height=f.kitty.height;const E=new ObjectGroup("robot");E.addObject(w),E.addObject(v),e.addLayer(E);const _=new ObjectGroup("callouts");for(let t of u.callouts.callouts){const r=new MapObject;r.shape=MapObject.Text,r.pos.x=(t.x+.2)*e.tileWidth,r.pos.y=(t.y+.2)*e.tileHeight,r.text=t.text,r.height=e.tileHeight,r.width=t.text.length*(r.font.pixelSize/2),r.wordWrap=!0,_.addObject(r)}return e.addLayer(_),e},e.tiledToBuf=function(t){var e,r,n,o,l,f;const u=null===(e=t.property("name"))||void 0===e?void 0:e.toString();if(!u)throw new Error("Could not determine level name. Please use map property 'name'");(0,s.fixFlippedCells)(t),tiled.log(`Now exporting level ${u}`);const h=new i.Level({name:u});t.property("name")&&t.property("name")!==h.name&&tiled.alert(`Level name property ignored. Level name will be "${h.name}" as filename.`),h.footer.belt_speed=parseFloat((null===(r=t.property("conveyor_speed"))||void 0===r?void 0:r.toString())||"0.6"),h.footer.music[0]=(null===(n=t.property("music_red"))||void 0===n?void 0:n.toString())||"",h.footer.music[1]=(null===(o=t.property("music_green"))||void 0===o?void 0:o.toString())||"",h.footer.music[2]=(null===(l=t.property("music_blue"))||void 0===l?void 0:l.toString())||"",null===(f=t.property("tags"))||void 0===f||f.toString().split(",").map((t=>{t&&h.tags.set_tag(t.trim().toLowerCase())}));const c=a(t);tiled.log(`Map size ${c.maxx-c.minx},${c.maxy-c.miny}`),h.grid.setSize(c.maxx-c.minx,c.maxy-c.miny);for(let e=0;e<t.layerCount;e++){const r=t.layerAt(e);if(r.isGroupLayer)throw new Error("Goup Layers currently not supported");if(r.isImageLayer)throw new Error("Image Layers currently not supported");if(r.isTileLayer){const e=r;for(let i=c.minx;i<c.maxx;i++){const n=i-c.minx;for(let o=c.miny;o<c.maxy;o++){const s=o-c.miny,l=e.tileAt(i,o);if(!l)continue;const a=l.properties();if(a.kind){const e={x:t.tileWidth*n,y:t.tileHeight*s};if("robot"===a.kind)h.robot.x=e.x+t.tileWidth/2,h.robot.y=e.y+t.tileHeight/2;else{if("kitty"!==a.kind)return`Tile at ${i},${o} in layer ${r.name} has unknown kind ${a.kind}`;h.kitty.x=e.x+t.tileWidth/2,h.kitty.y=e.y+t.tileHeight/2}}else if(a.base){const t=parseInt(a.base.toString());h.grid.setCellBase(n,s,t)}else if(a.paint){const t=parseInt(a.paint.toString());h.grid.setCellPaint(n,s,t)}else{if(!a.map)throw new Error(`Unsupported tile at pos (${i},${o}) in layer ${r.name}`);{const t=parseInt(a.map.toString());h.grid.setMapCell(n,s,t)}}}}}else if(r.isObjectLayer){const e=r;for(let i of e.objects)if(i.tile){const e=i.tile.properties();if(!e.kind)throw new Error("Unsuppored tile in object layer. Only robot and kitty are OK.");const n={x:i.x-c.minx*t.tileWidth,y:i.y-c.miny*t.tileHeight};if("robot"===e.kind)h.robot.x=n.x+t.tileWidth/2,h.robot.y=n.y-t.tileHeight/2;else{if("kitty"!==e.kind)return`Tile at ${i.x},${i.y} in layer ${r.name} has unknown kind ${e.kind}`;h.kitty.x=n.x+t.tileWidth/2,h.kitty.y=n.y-t.tileHeight/2}}else{if(!i.text)throw new Error(`Unsuppoted object in layer ${r.name}. Only tile and text are supported.`);{const e=Math.round(i.x/t.tileWidth)-c.minx,r=Math.round(i.y/t.tileHeight)-c.miny;h.callouts.addCallout(e,r,i.text)}}}}const d=h.serialize();return tiled.log("Kitty Map export success."),d},e.getMapLimits=a},330:t=>{"use strict";t.exports=JSON.parse('{"name":"rwklevelfiles","version":"3.0.0","description":"Extract level files from RWK web version","scripts":{"pack":"npx webpack","build":"tsc","install":"npm run pack","start":"node build/app.js","start:dev":"ts-node src/app.ts"},"repository":{"type":"git","url":"git+https://github.com/MrMasterplan/rwklevelfiles.git"},"author":"mrmasterplan","license":"GPL-3.0-or-later","bugs":{"url":"https://github.com/mrmasterplan/rwklevelfiles/issues"},"homepage":"https://github.com/mrmasterplan/rwklevelfiles","devDependencies":{"@types/glob":"*","@types/node":"*","ts-node":"*","ts-node-dev":"*","typescript":"*"},"dependencies":{"@mapeditor/tiled-api":"*","bitwise":"*","buffer":"*","ts-loader":"*","webpack":"*","webpack-cli":"*"}}')}},e={};function r(i){var n=e[i];if(void 0!==n)return n.exports;var o=e[i]={exports:{}};return t[i].call(o.exports,o,o.exports,r),o.exports}r.d=(t,e)=>{for(var i in e)r.o(e,i)&&!r.o(t,i)&&Object.defineProperty(t,i,{enumerable:!0,get:e[i]})},r.o=(t,e)=>Object.prototype.hasOwnProperty.call(t,e),r.r=t=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},r(73)})();
