+++ rwk/tilesets/robot.json
@@ -1,0 +1,42 @@
+{ "columns":0,
+ "grid":
+    {
+     "height":1,
+     "orientation":"orthogonal",
+     "width":1
+    },
+ "margin":0,
+ "name":"robot",
+ "spacing":0,
+ "tilecount":2,
+ "tiledversion":"1.5.0",
+ "tileheight":51,
+ "tiles":[
+        {
+         "id":0,
+         "image":"..\/tiles\/robot.png",
+         "imageheight":51,
+         "imagewidth":39,
+         "properties":[
+                {
+                 "name":"kind",
+                 "type":"string",
+                 "value":"robot"
+                }]
+        }, 
+        {
+         "id":1,
+         "image":"..\/tiles\/kitty.png",
+         "imageheight":28,
+         "imagewidth":44,
+         "properties":[
+                {
+                 "name":"kind",
+                 "type":"string",
+                 "value":"kitty"
+                }]
+        }],
+ "tilewidth":44,
+ "type":"tileset",
+ "version":1.5
+}
