+++ tiled/kitty.ts
@@ -1,10 +1,14 @@
 
 import * as kittypaint from "./kittypaint";
 import * as dotkitty from "./dotkitty";
+import * as dotkittyjs from "./dotkittyjs";
+import * as dotkittyb64 from "./dotkittyb64";
 import * as kittymenu from "./kittymenu";
 
 tiled.log("rwklevelfiles by MrMasterplan extensions for Tiled")
 tiled.log(__filename)
 kittypaint.register()
 dotkitty.register()
+dotkittyjs.register()
+dotkittyb64.register()
 kittymenu.register()
